# Conflicts:
#	Content.Client/Entry/IgnoredComponents.cs
#	Resources/Prototypes/Entities/Structures/Storage/Crates/base.yml

+++ Content.Tests/Shared/Chemistry/FixedPoint2_Tests.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using Content.Shared.Chemistry;
-using Content.Shared.Chemistry.Reagent;
 using Content.Shared.FixedPoint;
 using NUnit.Framework;
 
@@ -108,7 +106,7 @@ namespace Content.Tests.Shared.Chemistry
         [TestCase(2.005f, 201)]
         public void FloatRoundingTest(float a, int expected)
         {
-            var result = (int) MathF.Round(a * (float) MathF.Pow(10, 2), MidpointRounding.AwayFromZero);
+            var result = (int) MathF.Round(a * MathF.Pow(10, 2), MidpointRounding.AwayFromZero);
             Assert.That(result, Is.EqualTo(expected));
         }
